Progress on splitting
